feat: introduce new avenue of syncing which leverages json urls
During a full sync of a Saved history, it was found that the process was
taking extremely long due to the rate limit of the Reddit API. This is
due to all the calls executed for `more` comments in each Post's comment
tree.

So, instead of using the API directly, leverage the .json URLs of each
Post to retrieve the Post data and comments as-is; meaning no `more`
comments are also retrieved dynamically.

The local API will offer an endpoint later on initiate syncing all
Comments, including `more`, for a specific Post. This will also be
implemented as a Command so it may be executed by command line and cron
job.