fix: verify audio url for reddit-hosted videos
In certain cases, such as
/r/learnjavascript/comments/lht7ap/master_flexbox_2021_build_5_responsive/,
it has been found that though the is_gif data-point is set to false on
Reddit's side, there is still no audio file associated to the video
file. In such a case, it's warranted that the logic validate the
expected Audio URL prior to setting the Media Asset's property for that
data-point.