refactor: change default sync behavior to exclude comments
This is done to make syncing all Contents, from Saved items for example,
more streamlined. Also, it may not be the case in all syncs that the
associated Comments are needed or necessary; this will be deferred to
explicit action by the user to pull down the Comments associated to a
Post.