fix: do not attempt to reference video asset for removed content
It has been found that if a Post had a Video Media Asset that was
removed by Reddit (for copyright reasons, for example), the
Denormalization logic will still attempt to reference the video,
resulting in an error. To address, update the logic to explicitly check
for the existence of the Video.