fix: update mime type header detection logic to accommodate html
In some cases, a .gifv URL renders an HTML page instead of a media
source such as a .gifv, .gif, or .mp4 file. As a result, attempting to
detect the correct MIME Type raises an error. The detection logic must
be updated to include an HTML type, and in those cases, attempt to
reference and download the source's .mp4 link instead.