fix: add missing vars to .env to address failing cronjob
The full sync Command is failing due to the cron not having access to
the required Environment variables such as REDDIT_USERNAME. Though the
variables are set within the image environment, this is not accessible
from the cron's context. To address, add the required variables
explicitly to the .env file during its initialization within the
entrypoint execution.