fix: pull in missing gallery data from parent posts on crosspost syncs
A Post which has been cross-posted and thus has a Crosspost parent fails
to sync if the Crosspost parent has an Image Gallery because that data
is contained only within the Crosspost data, not within the target Post.
As a result, the current logic fails to detect a `Type` for the Post.

Implement a fix by pulling in the necessary Image Gallery data from the Crosspost
into the target Post.