fix: allow has_replies column to be null to serve as post-check flag
Using the Item Info endpoint on the Reddit API, it is not determinable
if a Comment has replies or not as the data-point is always set to an
empty string in the response for that endpoint. As a result, make the
`has_replies` column nullable to serve as a check to see if replies has
been attempted to sync for a specific Comment. If no replies were found,
the column should be false for that record; true if replies were found.